logisticai/services/inference/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
redis[asyncio]==5.0.4
python-dotenv==1.0.1
PyJWT==2.8.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
